Add source files -
26-Jan-18
These files are placed in a static AWS S3 bucket and serve from a URI domain. 
The code runs and in debug mode, the web worker is downloaded and executed.  
In the worker, the event.data is calculated, returned and assigned to the 'output' ID in the <div> element in the HTML file